+++ packages/flipCard/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/flipCard';
+export { default as FlipCard } from '../../../src/components/flipCard';
+export type * from '../../../src/components/flipCard';