+++ packages/nineGrid/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/nineGrid';
+export { default as NineGrid } from '../../../src/components/nineGrid';
+export type * from '../../../src/components/nineGrid';